Reorganize galaxy.tool_util.verify package/types for reuse outside module.
Add explicit exports to package galaxy.tool_util.verify.
