+++ lib/galaxy/tool_util/verify/__init__.py
@@ -48,6 +48,10 @@ from galaxy.tool_util.parser.util import (
 from galaxy.tool_util.parser.yaml import to_test_assert_list
 from galaxy.util import unicodify
 from galaxy.util.compression_utils import get_fileobj
+from ._types import (
+    ExpandedToolInputsJsonified,
+    ToolTestDescriptionDict,
+)
 from .asserts import verify_assertions
 from .test_data import TestDataResolver
 
@@ -641,3 +645,14 @@ def verify_file_contents_against_dict(
         keep_outputs_dir=test_data_target_dir,
         verify_extra_files=None,
     )
+
+
+__all__ = [
+    "DEFAULT_TEST_DATA_RESOLVER",
+    "ExpandedToolInputsJsonified",
+    "GetFilenameT",
+    "GetLocationT",
+    "ToolTestDescriptionDict",
+    "verify",
+    "verify_file_contents_against_dict",
+]
