+++ lib/galaxy/tool_util/verify/_types.py
@@ -4,7 +4,19 @@ from typing import (
     Any,
     Dict,
     List,
+    Optional,
     Tuple,
+)
+
+from typing_extensions import (
+    NotRequired,
+    TypedDict,
+)
+
+from galaxy.tool_util.parser.interface import (
+    AssertionList,
+    TestSourceTestOutputColllection,
+    ToolSourceTestOutputs,
 )
 
 # inputs that have been processed with parse.py and expanded out
@@ -16,3 +28,27 @@ RequiredFileTuple = Tuple[str, ExtraFileInfoDictT]
 RequiredFilesT = List[RequiredFileTuple]
 RequiredDataTablesT = List[str]
 RequiredLocFileT = List[str]
+
+
+class ToolTestDescriptionDict(TypedDict):
+    tool_id: str
+    tool_version: Optional[str]
+    name: str
+    test_index: int
+    inputs: ExpandedToolInputsJsonified
+    outputs: ToolSourceTestOutputs
+    output_collections: List[TestSourceTestOutputColllection]
+    stdout: Optional[AssertionList]
+    stderr: Optional[AssertionList]
+    expect_exit_code: Optional[int]
+    expect_failure: bool
+    expect_test_failure: bool
+    num_outputs: Optional[int]
+    command_line: Optional[AssertionList]
+    command_version: Optional[AssertionList]
+    required_files: List[Any]
+    required_data_tables: List[Any]
+    required_loc_files: List[str]
+    error: bool
+    exception: Optional[str]
+    maxseconds: NotRequired[Optional[int]]
