+++ lib/galaxy/webapps/galaxy/api/tools.py
@@ -31,7 +31,7 @@ from galaxy.schema.fetch_data import (
     FetchDataFormPayload,
     FetchDataPayload,
 )
-from galaxy.tool_util.verify.interactor import ToolTestDescriptionDict
+from galaxy.tool_util.verify import ToolTestDescriptionDict
 from galaxy.tools.evaluation import global_tool_errors
 from galaxy.util.zipstream import ZipstreamWrapper
 from galaxy.web import (
